+++ packages/compose/src/hooks/use-disabled/index.ts
@@ -45,14 +45,14 @@ export default function useDisabled( {
 			}
 
 			/** A variable keeping track of the previous updates in order to restore them. */
-			const updates: ( () => void )[] = [];
+			const updates: Function[] = [];
 			const disable = () => {
 				node.childNodes.forEach( ( child ) => {
 					if ( ! ( child instanceof defaultView.HTMLElement ) ) {
 						return;
 					}
-					if ( ! child.hasAttribute( 'inert' ) ) {
-						child.setAttribute( 'inert', '' );
+					if ( ! child.getAttribute( 'inert' ) ) {
+						child.setAttribute( 'inert', 'true' );
 						updates.push( () => {
 							child.removeAttribute( 'inert' );
 						} );